douban/requirements.txt
pycryptodome
requests
xxhash
